extract `help` into it's own subproject to minify `core` dependencies
asciitable brings in quite a few libraries including antlr... better
have the option to not use it, or at least structure it like that

```
+-de.vandermeer:asciitable:0.3.2
[info]   | +-de.vandermeer:ascii-utf-themes:0.0.1
[info]   | | +-de.vandermeer:skb-interfaces:0.0.1
[info]   | |   +-org.antlr:ST4:4.0.8
[info]   | |   | +-org.antlr:antlr-runtime:3.5.2
[info]   | |   |
[info]   | |   +-org.antlr:antlr4:4.5.1
[info]   | |   +-org.apache.commons:commons-lang3:3.4
[info]   | |   +-org.slf4j:slf4j-api:1.7.12 (evicted by: 2.0.7)
[info]   | |   +-org.slf4j:slf4j-api:2.0.7
[info]   | |
[info]   | +-de.vandermeer:char-translation:0.0.2
[info]   |   +-de.vandermeer:skb-interfaces:0.0.1
[info]   |     +-org.antlr:ST4:4.0.8
[info]   |     | +-org.antlr:antlr-runtime:3.5.2
[info]   |     |
[info]   |     +-org.antlr:antlr4:4.5.1
[info]   |     +-org.apache.commons:commons-lang3:3.4
[info]   |     +-org.slf4j:slf4j-api:1.7.12 (evicted by: 2.0.7)
[info]   |     +-org.slf4j:slf4j-api:2.0.7
[info]   |
```
