+++ test-schemas-domain-classes/src/main/scala/testdomains/generic/language.scala
@@ -6,6 +6,7 @@ trait language
     with traversals.ConcreteStoredConversions
     with neighboraccessors.Conversions
     with flatgraph.traversal.language
+    with flatgraph.help.language
     with flatgraph.Implicits {
   implicit def toGeneratedNodeStarters(domain: GenericDomain): GenericDomainNodeStarters = GenericDomainNodeStarters(domain)
 }
