+++ test-schemas-domain-classes/src/main/scala/testdomains/gratefuldead/language.scala
@@ -6,6 +6,7 @@ trait language
     with traversals.ConcreteStoredConversions
     with neighboraccessors.Conversions
     with flatgraph.traversal.language
+    with flatgraph.help.language
     with flatgraph.Implicits {
   implicit def toGeneratedNodeStarters(domain: GratefulDead): GratefulDeadNodeStarters = GratefulDeadNodeStarters(domain)
 }
